Unificato il modello per tutte le impostazioni
