+++ src/model/widgets/__init__.py
@@ -1,0 +1,6 @@
+from .settings_model import SettingsModel
+
+
+__all__ = [
+    "SettingsModel"
+]
