--- src/model/widgets/settings/__init__.py
@@ -1,9 +1,0 @@
-from .set_path_model import SetPathModel
-from .set_policy_model import SetPolicyModel
-from .set_quota_disk_model import SetQuotaDiskModel
-
-__all__ = [
-    "SetPathModel",
-    "SetPolicyModel",
-    "SetQuotaDiskModel"
-]
